backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
nexus-rpc==1.3.0
protobuf==6.33.6
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
SQLAlchemy==2.0.48
starlette==0.52.1
supabase==2.28.3
temporalio==1.23.0
types-protobuf==6.32.1.20260221
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
